src/web/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic>=2.11.7,<3.0.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
pyyaml==6.0.1
psycopg2-binary==2.9.9